boost调试

Created-by: xiaoxinglxs
Commit-by: xiaoxinglxs
Merged-by: xiaoxinglxs
Description: boost调试

See merge request: openUBMC/libmcpp!168

+++ conanfile.py
@@ -30,6 +30,8 @@ class AppConan(ConanBase):
         d.cpp_args.append("-Wno-pedantic")
         d.cpp_args.append("-fno-strict-aliasing")
         d.cpp_args.append("-Wno-deprecated-copy")
+        # 增加boost包lib路径到链接参数 待2025.7.30 bmc_sdk发布正式包后删除
+        d.cpp_link_args.append("-L/root/.conan/data/boost/1.82.0.B001/openUBMC.release/rc/package/295f5ceaff90a1afe2a22ca78ccdeb749ab95b30/lib")
         d.generate()
 
         os.environ["PKG_CONFIG"] = "/usr/bin/pkg-config"
@@ -88,9 +90,6 @@ class AppConan(ConanBase):
         print(f"[WARNING] src_lib_dir 目录: {src_lib_dir}")
         print(f"[WARNING] dst_boost_dir 目录: {dst_boost_dir}")
         # 先在test/boost/lib目录下生成无版本号so/a文件
-        current_dir2 = os.path.dirname(src_lib_dir)
-        print(f"[WARNING] test/boost/lib 所在目录: {current_dir2}")
-        print(f"[WARNING] {current_dir2} 下文件/文件夹: {os.listdir(current_dir2)}")
         so_src_local = os.path.join(src_lib_dir, "libboost_program_options.so.1.82.0")
         so_dst_local = os.path.join(src_lib_dir, "libboost_program_options.so")
         a_src_local = os.path.join(src_lib_dir, "libboost_program_options.a.1.82.0")
@@ -123,6 +122,9 @@ class AppConan(ConanBase):
                 for filename in os.listdir(src_lib_dir):
                     src_file = os.path.join(src_lib_dir, filename)
                     dst_file = os.path.join(dst_boost_dir, filename)
+                    if os.path.exists(dst_file):
+                        print(f"[INFO] 目标文件已存在，跳过拷贝: {dst_file}")
+                        continue
                     shutil.copy2(src_file, dst_file)
                     print(f"[INFO] 已拷贝 {src_file} 到 {dst_file}")
             except Exception as e:
